refactor: rename AccordionListPage to AccordionPage
Trim the one over-stacked component name — 'List' was redundant since the
accordion sections already imply it. Rename the file (history preserved) and
update the two view references. Behaviour and rendered output unchanged.
