+++ src/components/AccordionPage.vue
